fix(gchat): move GoogleChatAdapterConfig to ./types for konsistent
